Refactor IssueCreate.js and IssueUpdate/component.json
Update IssueCreate.js to return the data received from httpRequest() instead of an empty object.

Update component.json files to change the type of "issue_id" property from integer to string.

Update openapi.yml to add required properties for creating an issue and update the response schema for creating an issue.

Update openapi.yml to remove the "issue" property from the required properties for updating an issue.

Update openapi.yml to change the response status code for updating an issue from 201 to 204.